requirements.txt
feedparser==6.0.11
html2text==2024.2.26
python-dotenv==1.0.1
python-telegram-bot==21.3
Requests==2.32.3